+++ .cache/60/d70d2aed01704e2468f8bfac01b1a4.json
@@ -1,0 +1,1 @@
+{"id":"node_modules/axios/lib/core/createError.js","dependencies":[{"name":"/home/edycakra/Documents/H8/P2/Porto/Kanban/client/kanban-porto/package.json","includedInParent":true,"mtime":1581654140184},{"name":"/home/edycakra/Documents/H8/P2/Porto/Kanban/client/kanban-porto/node_modules/axios/package.json","includedInParent":true,"mtime":1581653445411},{"name":"./enhanceError","loc":{"line":3,"column":27},"parent":"/home/edycakra/Documents/H8/P2/Porto/Kanban/client/kanban-porto/node_modules/axios/lib/core/createError.js","resolved":"/home/edycakra/Documents/H8/P2/Porto/Kanban/client/kanban-porto/node_modules/axios/lib/core/enhanceError.js"}],"generated":{"js":"'use strict';\n\nvar enhanceError = require('./enhanceError');\n\n/**\n * Create an Error with the specified message, config, error code, request and response.\n *\n * @param {string} message The error message.\n * @param {Object} config The config.\n * @param {string} [code] The error code (for example, 'ECONNABORTED').\n * @param {Object} [request] The request.\n * @param {Object} [response] The response.\n * @returns {Error} The created error.\n */\nmodule.exports = function createError(message, config, code, request, response) {\n  var error = new Error(message);\n  return enhanceError(error, config, code, request, response);\n};\n"},"sourceMaps":{"js":{"mappings":[{"source":"node_modules/axios/lib/core/createError.js","original":{"line":1,"column":0},"generated":{"line":1,"column":0}},{"source":"node_modules/axios/lib/core/createError.js","original":{"line":2,"column":0},"generated":{"line":2,"column":0}},{"source":"node_modules/axios/lib/core/createError.js","original":{"line":3,"column":0},"generated":{"line":3,"column":0}},{"source":"node_modules/axios/lib/core/createError.js","original":{"line":4,"column":0},"generated":{"line":4,"column":0}},{"source":"node_modules/axios/lib/core/createError.js","original":{"line":5,"column":0},"generated":{"line":5,"column":0}},{"source":"node_modules/axios/lib/core/createError.js","original":{"line":6,"column":0},"generated":{"line":6,"column":0}},{"source":"node_modules/axios/lib/core/createError.js","original":{"line":7,"column":0},"generated":{"line":7,"column":0}},{"source":"node_modules/axios/lib/core/createError.js","original":{"line":8,"column":0},"generated":{"line":8,"column":0}},{"source":"node_modules/axios/lib/core/createError.js","original":{"line":9,"column":0},"generated":{"line":9,"column":0}},{"source":"node_modules/axios/lib/core/createError.js","original":{"line":10,"column":0},"generated":{"line":10,"column":0}},{"source":"node_modules/axios/lib/core/createError.js","original":{"line":11,"column":0},"generated":{"line":11,"column":0}},{"source":"node_modules/axios/lib/core/createError.js","original":{"line":12,"column":0},"generated":{"line":12,"column":0}},{"source":"node_modules/axios/lib/core/createError.js","original":{"line":13,"column":0},"generated":{"line":13,"column":0}},{"source":"node_modules/axios/lib/core/createError.js","original":{"line":14,"column":0},"generated":{"line":14,"column":0}},{"source":"node_modules/axios/lib/core/createError.js","original":{"line":15,"column":0},"generated":{"line":15,"column":0}},{"source":"node_modules/axios/lib/core/createError.js","original":{"line":16,"column":0},"generated":{"line":16,"column":0}},{"source":"node_modules/axios/lib/core/createError.js","original":{"line":17,"column":0},"generated":{"line":17,"column":0}},{"source":"node_modules/axios/lib/core/createError.js","original":{"line":18,"column":0},"generated":{"line":18,"column":0}},{"source":"node_modules/axios/lib/core/createError.js","original":{"line":19,"column":0},"generated":{"line":19,"column":0}}],"sources":{"node_modules/axios/lib/core/createError.js":"'use strict';\n\nvar enhanceError = require('./enhanceError');\n\n/**\n * Create an Error with the specified message, config, error code, request and response.\n *\n * @param {string} message The error message.\n * @param {Object} config The config.\n * @param {string} [code] The error code (for example, 'ECONNABORTED').\n * @param {Object} [request] The request.\n * @param {Object} [response] The response.\n * @returns {Error} The created error.\n */\nmodule.exports = function createError(message, config, code, request, response) {\n  var error = new Error(message);\n  return enhanceError(error, config, code, request, response);\n};\n"},"lineCount":19}},"error":null,"hash":"da76eeae4052065cdff865fd065e585f","cacheData":{"env":{}}}
