+++ .cache/89/6917c11dacca4e0464ba3d5f0810fd.json
@@ -1,1 +1,1 @@
-{"id":"index.html","dependencies":[{"name":"/home/edycakra/Documents/H8/P2/Porto/Kanban/client/kanban-porto/src/main.js","dynamic":true,"resolved":"/home/edycakra/Documents/H8/P2/Porto/Kanban/client/kanban-porto/src/main.js","parent":"/home/edycakra/Documents/H8/P2/Porto/Kanban/client/kanban-porto/index.html"}],"generated":{"html":"<!DOCTYPE html>\n<html lang=\"en\">\n<head>\n    <meta charset=\"UTF-8\">\n    <meta name=\"viewport\" content=\"width=device-width, initial-scale=1.0\">\n    <meta http-equiv=\"X-UA-Compatible\" content=\"ie=edge\">\n    <!-- <meta name=\"google-signin-client_id\" content=\"777356335430-brvkhqdho2rei5dtss0rqe8c8lmug1gl.apps.googleusercontent.com\"> -->\n    <title>KANBAN</title>\n    <!-- vue  -->\n    <script src=\"https://cdn.jsdelivr.net/npm/vue/dist/vue.js\"></script>\n    <!-- bulma  -->\n    <link rel=\"stylesheet\" href=\"https://cdn.jsdelivr.net/npm/bulma@0.8.0/css/bulma.min.css\">\n    <!-- fontawesome  -->\n    <script defer=\"\" src=\"https://use.fontawesome.com/releases/v5.3.1/js/all.js\"></script>\n</head>\n<body>\n    <!-- vue  -->\n    <div id=\"app\"></div>\n    <!-- script vue-->\n    <script src=\"/1c90ff38a08209f9ebd4d05d1e43358e.js\"></script>\n    <script src=\"https://unpkg.com/axios/dist/axios.min.js\"></script>\n    <!-- <script src=\"https://apis.google.com/js/platform.js\" async defer></script> -->\n</body>\n</html>"},"sourceMaps":null,"error":null,"hash":"c28ed7eeeb52585d316bf5168143a96e","cacheData":{}}
+{"id":"index.html","dependencies":[{"name":"/home/edycakra/Documents/H8/P2/Porto/Kanban/client/kanban-porto/src/main.js","dynamic":true,"resolved":"/home/edycakra/Documents/H8/P2/Porto/Kanban/client/kanban-porto/src/main.js","parent":"/home/edycakra/Documents/H8/P2/Porto/Kanban/client/kanban-porto/index.html"}],"generated":{"html":"<!DOCTYPE html>\n<html lang=\"en\">\n<head>\n    <meta charset=\"UTF-8\">\n    <meta name=\"viewport\" content=\"width=device-width, initial-scale=1.0\">\n    <meta http-equiv=\"X-UA-Compatible\" content=\"ie=edge\">\n    <!-- <meta name=\"google-signin-client_id\" content=\"777356335430-brvkhqdho2rei5dtss0rqe8c8lmug1gl.apps.googleusercontent.com\"> -->\n    <title>KANBAN</title>\n    <!-- vue  -->\n    <script src=\"https://cdn.jsdelivr.net/npm/vue/dist/vue.js\"></script>\n    <!-- bulma  -->\n    <link rel=\"stylesheet\" href=\"https://cdn.jsdelivr.net/npm/bulma@0.8.0/css/bulma.min.css\">\n    <script defer=\"\" src=\"https://use.fontawesome.com/releases/v5.3.1/js/all.js\"></script>\n</head>\n<body>\n    <div id=\"app\"></div>\n    <script src=\"/1c90ff38a08209f9ebd4d05d1e43358e.js\"></script>\n    <script src=\"https://unpkg.com/axios/dist/axios.min.js\"></script>\n    <!-- <script src=\"https://apis.google.com/js/platform.js\" async defer></script> -->\n</body>\n</html>"},"sourceMaps":null,"error":null,"hash":"526ca4556e668a713f38099adc985d49","cacheData":{}}
