+++ .cache/9c/45a4e3fd4176821f939e83fc922d8c.json
@@ -1,0 +1,1 @@
+{"id":"node_modules/axios/lib/helpers/bind.js","dependencies":[{"name":"/home/edycakra/Documents/H8/P2/Porto/Kanban/client/kanban-porto/package.json","includedInParent":true,"mtime":1581654140184},{"name":"/home/edycakra/Documents/H8/P2/Porto/Kanban/client/kanban-porto/node_modules/axios/package.json","includedInParent":true,"mtime":1581653445411}],"generated":{"js":"'use strict';\n\nmodule.exports = function bind(fn, thisArg) {\n  return function wrap() {\n    var args = new Array(arguments.length);\n    for (var i = 0; i < args.length; i++) {\n      args[i] = arguments[i];\n    }\n    return fn.apply(thisArg, args);\n  };\n};\n"},"sourceMaps":{"js":{"mappings":[{"source":"node_modules/axios/lib/helpers/bind.js","original":{"line":1,"column":0},"generated":{"line":1,"column":0}},{"source":"node_modules/axios/lib/helpers/bind.js","original":{"line":2,"column":0},"generated":{"line":2,"column":0}},{"source":"node_modules/axios/lib/helpers/bind.js","original":{"line":3,"column":0},"generated":{"line":3,"column":0}},{"source":"node_modules/axios/lib/helpers/bind.js","original":{"line":4,"column":0},"generated":{"line":4,"column":0}},{"source":"node_modules/axios/lib/helpers/bind.js","original":{"line":5,"column":0},"generated":{"line":5,"column":0}},{"source":"node_modules/axios/lib/helpers/bind.js","original":{"line":6,"column":0},"generated":{"line":6,"column":0}},{"source":"node_modules/axios/lib/helpers/bind.js","original":{"line":7,"column":0},"generated":{"line":7,"column":0}},{"source":"node_modules/axios/lib/helpers/bind.js","original":{"line":8,"column":0},"generated":{"line":8,"column":0}},{"source":"node_modules/axios/lib/helpers/bind.js","original":{"line":9,"column":0},"generated":{"line":9,"column":0}},{"source":"node_modules/axios/lib/helpers/bind.js","original":{"line":10,"column":0},"generated":{"line":10,"column":0}},{"source":"node_modules/axios/lib/helpers/bind.js","original":{"line":11,"column":0},"generated":{"line":11,"column":0}},{"source":"node_modules/axios/lib/helpers/bind.js","original":{"line":12,"column":0},"generated":{"line":12,"column":0}}],"sources":{"node_modules/axios/lib/helpers/bind.js":"'use strict';\n\nmodule.exports = function bind(fn, thisArg) {\n  return function wrap() {\n    var args = new Array(arguments.length);\n    for (var i = 0; i < args.length; i++) {\n      args[i] = arguments[i];\n    }\n    return fn.apply(thisArg, args);\n  };\n};\n"},"lineCount":12}},"error":null,"hash":"3a23a80cb8e569d4fb5aae215acffdbb","cacheData":{"env":{}}}
