+++ .cache/a2/303ea6632d26013cfb2bf21fab5191.json
@@ -1,0 +1,1 @@
+{"id":"node_modules/axios/lib/helpers/isAbsoluteURL.js","dependencies":[{"name":"/home/edycakra/Documents/H8/P2/Porto/Kanban/client/kanban-porto/package.json","includedInParent":true,"mtime":1581654140184},{"name":"/home/edycakra/Documents/H8/P2/Porto/Kanban/client/kanban-porto/node_modules/axios/package.json","includedInParent":true,"mtime":1581653445411}],"generated":{"js":"'use strict';\n\n/**\n * Determines whether the specified URL is absolute\n *\n * @param {string} url The URL to test\n * @returns {boolean} True if the specified URL is absolute, otherwise false\n */\nmodule.exports = function isAbsoluteURL(url) {\n  // A URL is considered absolute if it begins with \"<scheme>://\" or \"//\" (protocol-relative URL).\n  // RFC 3986 defines scheme name as a sequence of characters beginning with a letter and followed\n  // by any combination of letters, digits, plus, period, or hyphen.\n  return /^([a-z][a-z\\d\\+\\-\\.]*:)?\\/\\//i.test(url);\n};\n"},"sourceMaps":{"js":{"mappings":[{"source":"node_modules/axios/lib/helpers/isAbsoluteURL.js","original":{"line":1,"column":0},"generated":{"line":1,"column":0}},{"source":"node_modules/axios/lib/helpers/isAbsoluteURL.js","original":{"line":2,"column":0},"generated":{"line":2,"column":0}},{"source":"node_modules/axios/lib/helpers/isAbsoluteURL.js","original":{"line":3,"column":0},"generated":{"line":3,"column":0}},{"source":"node_modules/axios/lib/helpers/isAbsoluteURL.js","original":{"line":4,"column":0},"generated":{"line":4,"column":0}},{"source":"node_modules/axios/lib/helpers/isAbsoluteURL.js","original":{"line":5,"column":0},"generated":{"line":5,"column":0}},{"source":"node_modules/axios/lib/helpers/isAbsoluteURL.js","original":{"line":6,"column":0},"generated":{"line":6,"column":0}},{"source":"node_modules/axios/lib/helpers/isAbsoluteURL.js","original":{"line":7,"column":0},"generated":{"line":7,"column":0}},{"source":"node_modules/axios/lib/helpers/isAbsoluteURL.js","original":{"line":8,"column":0},"generated":{"line":8,"column":0}},{"source":"node_modules/axios/lib/helpers/isAbsoluteURL.js","original":{"line":9,"column":0},"generated":{"line":9,"column":0}},{"source":"node_modules/axios/lib/helpers/isAbsoluteURL.js","original":{"line":10,"column":0},"generated":{"line":10,"column":0}},{"source":"node_modules/axios/lib/helpers/isAbsoluteURL.js","original":{"line":11,"column":0},"generated":{"line":11,"column":0}},{"source":"node_modules/axios/lib/helpers/isAbsoluteURL.js","original":{"line":12,"column":0},"generated":{"line":12,"column":0}},{"source":"node_modules/axios/lib/helpers/isAbsoluteURL.js","original":{"line":13,"column":0},"generated":{"line":13,"column":0}},{"source":"node_modules/axios/lib/helpers/isAbsoluteURL.js","original":{"line":14,"column":0},"generated":{"line":14,"column":0}},{"source":"node_modules/axios/lib/helpers/isAbsoluteURL.js","original":{"line":15,"column":0},"generated":{"line":15,"column":0}}],"sources":{"node_modules/axios/lib/helpers/isAbsoluteURL.js":"'use strict';\n\n/**\n * Determines whether the specified URL is absolute\n *\n * @param {string} url The URL to test\n * @returns {boolean} True if the specified URL is absolute, otherwise false\n */\nmodule.exports = function isAbsoluteURL(url) {\n  // A URL is considered absolute if it begins with \"<scheme>://\" or \"//\" (protocol-relative URL).\n  // RFC 3986 defines scheme name as a sequence of characters beginning with a letter and followed\n  // by any combination of letters, digits, plus, period, or hyphen.\n  return /^([a-z][a-z\\d\\+\\-\\.]*:)?\\/\\//i.test(url);\n};\n"},"lineCount":15}},"error":null,"hash":"899c22b0645c747aa2fc667f528e7b0e","cacheData":{"env":{}}}
