+++ .cache/c1/d42f3ee403c787871c7261e549ba24.json
@@ -1,0 +1,1 @@
+{"id":"node_modules/axios/index.js","dependencies":[{"name":"/home/edycakra/Documents/H8/P2/Porto/Kanban/client/kanban-porto/package.json","includedInParent":true,"mtime":1581654140184},{"name":"/home/edycakra/Documents/H8/P2/Porto/Kanban/client/kanban-porto/node_modules/axios/package.json","includedInParent":true,"mtime":1581653445411},{"name":"./lib/axios","loc":{"line":1,"column":25},"parent":"/home/edycakra/Documents/H8/P2/Porto/Kanban/client/kanban-porto/node_modules/axios/index.js","resolved":"/home/edycakra/Documents/H8/P2/Porto/Kanban/client/kanban-porto/node_modules/axios/lib/axios.js"}],"generated":{"js":"module.exports = require('./lib/axios');"},"sourceMaps":{"js":{"mappings":[{"source":"node_modules/axios/index.js","original":{"line":1,"column":0},"generated":{"line":1,"column":0}}],"sources":{"node_modules/axios/index.js":"module.exports = require('./lib/axios');"},"lineCount":1}},"error":null,"hash":"ab539cfba3ff8f694b70f9e368216c3d","cacheData":{"env":{}}}
