+++ .cache/d4/77f5b10d7b7da44af87884ff632507.json
@@ -1,0 +1,1 @@
+{"id":"node_modules/axios/lib/core/buildFullPath.js","dependencies":[{"name":"/home/edycakra/Documents/H8/P2/Porto/Kanban/client/kanban-porto/package.json","includedInParent":true,"mtime":1581654140184},{"name":"/home/edycakra/Documents/H8/P2/Porto/Kanban/client/kanban-porto/node_modules/axios/package.json","includedInParent":true,"mtime":1581653445411},{"name":"../helpers/isAbsoluteURL","loc":{"line":3,"column":28},"parent":"/home/edycakra/Documents/H8/P2/Porto/Kanban/client/kanban-porto/node_modules/axios/lib/core/buildFullPath.js","resolved":"/home/edycakra/Documents/H8/P2/Porto/Kanban/client/kanban-porto/node_modules/axios/lib/helpers/isAbsoluteURL.js"},{"name":"../helpers/combineURLs","loc":{"line":4,"column":26},"parent":"/home/edycakra/Documents/H8/P2/Porto/Kanban/client/kanban-porto/node_modules/axios/lib/core/buildFullPath.js","resolved":"/home/edycakra/Documents/H8/P2/Porto/Kanban/client/kanban-porto/node_modules/axios/lib/helpers/combineURLs.js"}],"generated":{"js":"'use strict';\n\nvar isAbsoluteURL = require('../helpers/isAbsoluteURL');\nvar combineURLs = require('../helpers/combineURLs');\n\n/**\n * Creates a new URL by combining the baseURL with the requestedURL,\n * only when the requestedURL is not already an absolute URL.\n * If the requestURL is absolute, this function returns the requestedURL untouched.\n *\n * @param {string} baseURL The base URL\n * @param {string} requestedURL Absolute or relative URL to combine\n * @returns {string} The combined full path\n */\nmodule.exports = function buildFullPath(baseURL, requestedURL) {\n  if (baseURL && !isAbsoluteURL(requestedURL)) {\n    return combineURLs(baseURL, requestedURL);\n  }\n  return requestedURL;\n};\n"},"sourceMaps":{"js":{"mappings":[{"source":"node_modules/axios/lib/core/buildFullPath.js","original":{"line":1,"column":0},"generated":{"line":1,"column":0}},{"source":"node_modules/axios/lib/core/buildFullPath.js","original":{"line":2,"column":0},"generated":{"line":2,"column":0}},{"source":"node_modules/axios/lib/core/buildFullPath.js","original":{"line":3,"column":0},"generated":{"line":3,"column":0}},{"source":"node_modules/axios/lib/core/buildFullPath.js","original":{"line":4,"column":0},"generated":{"line":4,"column":0}},{"source":"node_modules/axios/lib/core/buildFullPath.js","original":{"line":5,"column":0},"generated":{"line":5,"column":0}},{"source":"node_modules/axios/lib/core/buildFullPath.js","original":{"line":6,"column":0},"generated":{"line":6,"column":0}},{"source":"node_modules/axios/lib/core/buildFullPath.js","original":{"line":7,"column":0},"generated":{"line":7,"column":0}},{"source":"node_modules/axios/lib/core/buildFullPath.js","original":{"line":8,"column":0},"generated":{"line":8,"column":0}},{"source":"node_modules/axios/lib/core/buildFullPath.js","original":{"line":9,"column":0},"generated":{"line":9,"column":0}},{"source":"node_modules/axios/lib/core/buildFullPath.js","original":{"line":10,"column":0},"generated":{"line":10,"column":0}},{"source":"node_modules/axios/lib/core/buildFullPath.js","original":{"line":11,"column":0},"generated":{"line":11,"column":0}},{"source":"node_modules/axios/lib/core/buildFullPath.js","original":{"line":12,"column":0},"generated":{"line":12,"column":0}},{"source":"node_modules/axios/lib/core/buildFullPath.js","original":{"line":13,"column":0},"generated":{"line":13,"column":0}},{"source":"node_modules/axios/lib/core/buildFullPath.js","original":{"line":14,"column":0},"generated":{"line":14,"column":0}},{"source":"node_modules/axios/lib/core/buildFullPath.js","original":{"line":15,"column":0},"generated":{"line":15,"column":0}},{"source":"node_modules/axios/lib/core/buildFullPath.js","original":{"line":16,"column":0},"generated":{"line":16,"column":0}},{"source":"node_modules/axios/lib/core/buildFullPath.js","original":{"line":17,"column":0},"generated":{"line":17,"column":0}},{"source":"node_modules/axios/lib/core/buildFullPath.js","original":{"line":18,"column":0},"generated":{"line":18,"column":0}},{"source":"node_modules/axios/lib/core/buildFullPath.js","original":{"line":19,"column":0},"generated":{"line":19,"column":0}},{"source":"node_modules/axios/lib/core/buildFullPath.js","original":{"line":20,"column":0},"generated":{"line":20,"column":0}},{"source":"node_modules/axios/lib/core/buildFullPath.js","original":{"line":21,"column":0},"generated":{"line":21,"column":0}}],"sources":{"node_modules/axios/lib/core/buildFullPath.js":"'use strict';\n\nvar isAbsoluteURL = require('../helpers/isAbsoluteURL');\nvar combineURLs = require('../helpers/combineURLs');\n\n/**\n * Creates a new URL by combining the baseURL with the requestedURL,\n * only when the requestedURL is not already an absolute URL.\n * If the requestURL is absolute, this function returns the requestedURL untouched.\n *\n * @param {string} baseURL The base URL\n * @param {string} requestedURL Absolute or relative URL to combine\n * @returns {string} The combined full path\n */\nmodule.exports = function buildFullPath(baseURL, requestedURL) {\n  if (baseURL && !isAbsoluteURL(requestedURL)) {\n    return combineURLs(baseURL, requestedURL);\n  }\n  return requestedURL;\n};\n"},"lineCount":21}},"error":null,"hash":"ac2fbdb20575524c2da9e38c025926d2","cacheData":{"env":{}}}
