+++ .cache/f7/adc12b03c6291c694d25720b0cca8e.json
@@ -1,0 +1,1 @@
+{"id":"node_modules/axios/lib/core/transformData.js","dependencies":[{"name":"/home/edycakra/Documents/H8/P2/Porto/Kanban/client/kanban-porto/package.json","includedInParent":true,"mtime":1581654140184},{"name":"/home/edycakra/Documents/H8/P2/Porto/Kanban/client/kanban-porto/node_modules/axios/package.json","includedInParent":true,"mtime":1581653445411},{"name":"./../utils","loc":{"line":3,"column":20},"parent":"/home/edycakra/Documents/H8/P2/Porto/Kanban/client/kanban-porto/node_modules/axios/lib/core/transformData.js","resolved":"/home/edycakra/Documents/H8/P2/Porto/Kanban/client/kanban-porto/node_modules/axios/lib/utils.js"}],"generated":{"js":"'use strict';\n\nvar utils = require('./../utils');\n\n/**\n * Transform the data for a request or a response\n *\n * @param {Object|String} data The data to be transformed\n * @param {Array} headers The headers for the request or response\n * @param {Array|Function} fns A single function or Array of functions\n * @returns {*} The resulting transformed data\n */\nmodule.exports = function transformData(data, headers, fns) {\n  /*eslint no-param-reassign:0*/\n  utils.forEach(fns, function transform(fn) {\n    data = fn(data, headers);\n  });\n\n  return data;\n};\n"},"sourceMaps":{"js":{"mappings":[{"source":"node_modules/axios/lib/core/transformData.js","original":{"line":1,"column":0},"generated":{"line":1,"column":0}},{"source":"node_modules/axios/lib/core/transformData.js","original":{"line":2,"column":0},"generated":{"line":2,"column":0}},{"source":"node_modules/axios/lib/core/transformData.js","original":{"line":3,"column":0},"generated":{"line":3,"column":0}},{"source":"node_modules/axios/lib/core/transformData.js","original":{"line":4,"column":0},"generated":{"line":4,"column":0}},{"source":"node_modules/axios/lib/core/transformData.js","original":{"line":5,"column":0},"generated":{"line":5,"column":0}},{"source":"node_modules/axios/lib/core/transformData.js","original":{"line":6,"column":0},"generated":{"line":6,"column":0}},{"source":"node_modules/axios/lib/core/transformData.js","original":{"line":7,"column":0},"generated":{"line":7,"column":0}},{"source":"node_modules/axios/lib/core/transformData.js","original":{"line":8,"column":0},"generated":{"line":8,"column":0}},{"source":"node_modules/axios/lib/core/transformData.js","original":{"line":9,"column":0},"generated":{"line":9,"column":0}},{"source":"node_modules/axios/lib/core/transformData.js","original":{"line":10,"column":0},"generated":{"line":10,"column":0}},{"source":"node_modules/axios/lib/core/transformData.js","original":{"line":11,"column":0},"generated":{"line":11,"column":0}},{"source":"node_modules/axios/lib/core/transformData.js","original":{"line":12,"column":0},"generated":{"line":12,"column":0}},{"source":"node_modules/axios/lib/core/transformData.js","original":{"line":13,"column":0},"generated":{"line":13,"column":0}},{"source":"node_modules/axios/lib/core/transformData.js","original":{"line":14,"column":0},"generated":{"line":14,"column":0}},{"source":"node_modules/axios/lib/core/transformData.js","original":{"line":15,"column":0},"generated":{"line":15,"column":0}},{"source":"node_modules/axios/lib/core/transformData.js","original":{"line":16,"column":0},"generated":{"line":16,"column":0}},{"source":"node_modules/axios/lib/core/transformData.js","original":{"line":17,"column":0},"generated":{"line":17,"column":0}},{"source":"node_modules/axios/lib/core/transformData.js","original":{"line":18,"column":0},"generated":{"line":18,"column":0}},{"source":"node_modules/axios/lib/core/transformData.js","original":{"line":19,"column":0},"generated":{"line":19,"column":0}},{"source":"node_modules/axios/lib/core/transformData.js","original":{"line":20,"column":0},"generated":{"line":20,"column":0}},{"source":"node_modules/axios/lib/core/transformData.js","original":{"line":21,"column":0},"generated":{"line":21,"column":0}}],"sources":{"node_modules/axios/lib/core/transformData.js":"'use strict';\n\nvar utils = require('./../utils');\n\n/**\n * Transform the data for a request or a response\n *\n * @param {Object|String} data The data to be transformed\n * @param {Array} headers The headers for the request or response\n * @param {Array|Function} fns A single function or Array of functions\n * @returns {*} The resulting transformed data\n */\nmodule.exports = function transformData(data, headers, fns) {\n  /*eslint no-param-reassign:0*/\n  utils.forEach(fns, function transform(fn) {\n    data = fn(data, headers);\n  });\n\n  return data;\n};\n"},"lineCount":21}},"error":null,"hash":"68a8aed26f245cc7dbbe278b0c793b9b","cacheData":{"env":{}}}
