+++ .cache/fb/0dffb0fe32870dc8e2f4d9086ca49e.json
@@ -1,0 +1,1 @@
+{"id":"node_modules/axios/lib/cancel/isCancel.js","dependencies":[{"name":"/home/edycakra/Documents/H8/P2/Porto/Kanban/client/kanban-porto/package.json","includedInParent":true,"mtime":1581654140184},{"name":"/home/edycakra/Documents/H8/P2/Porto/Kanban/client/kanban-porto/node_modules/axios/package.json","includedInParent":true,"mtime":1581653445411}],"generated":{"js":"'use strict';\n\nmodule.exports = function isCancel(value) {\n  return !!(value && value.__CANCEL__);\n};\n"},"sourceMaps":{"js":{"mappings":[{"source":"node_modules/axios/lib/cancel/isCancel.js","original":{"line":1,"column":0},"generated":{"line":1,"column":0}},{"source":"node_modules/axios/lib/cancel/isCancel.js","original":{"line":2,"column":0},"generated":{"line":2,"column":0}},{"source":"node_modules/axios/lib/cancel/isCancel.js","original":{"line":3,"column":0},"generated":{"line":3,"column":0}},{"source":"node_modules/axios/lib/cancel/isCancel.js","original":{"line":4,"column":0},"generated":{"line":4,"column":0}},{"source":"node_modules/axios/lib/cancel/isCancel.js","original":{"line":5,"column":0},"generated":{"line":5,"column":0}},{"source":"node_modules/axios/lib/cancel/isCancel.js","original":{"line":6,"column":0},"generated":{"line":6,"column":0}}],"sources":{"node_modules/axios/lib/cancel/isCancel.js":"'use strict';\n\nmodule.exports = function isCancel(value) {\n  return !!(value && value.__CANCEL__);\n};\n"},"lineCount":6}},"error":null,"hash":"49d6c900dd16450dc9e91622fbb2de38","cacheData":{"env":{}}}
