+++ index.html
@@ -10,13 +10,10 @@
     <script src="https://cdn.jsdelivr.net/npm/vue/dist/vue.js"></script>
     <!-- bulma  -->
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bulma@0.8.0/css/bulma.min.css">
-    <!-- fontawesome  -->
     <script defer src="https://use.fontawesome.com/releases/v5.3.1/js/all.js"></script>
 </head>
 <body>
-    <!-- vue  -->
     <div id="app"></div>
-    <!-- script vue-->
     <script src="/src/main.js"></script>
     <script src="https://unpkg.com/axios/dist/axios.min.js"></script>
     <!-- <script src="https://apis.google.com/js/platform.js" async defer></script> -->
